fix: rename file and check enabled toggle on path
