+++ src/modules/bonus/index.ts
@@ -18,4 +18,4 @@ export {
   bonusOnboardingCarouselConfig,
   type BonusOnboardingScreenName,
 } from './onboarding';
-export {useIsBonusActiveForUser} from './use-is-bonus-active';
+export {useIsBonusActiveForUser} from './use-is-bonus-active-for-user';
